Move testcases so they are in appropriate places
