+++ rewrite-java-test/src/test/java/org/openrewrite/java/RemoveUnusedImportsTest.java
@@ -2035,4 +2035,37 @@ class RemoveUnusedImportsTest implements RewriteTest {
           )
         );
     }
+
+    @Test
+    void staticNestedInnerClass() {
+        // language=java
+        rewriteRun(
+          java(
+            """
+              package a;
+              
+              public class A {
+                  public final class B {}
+                  public static class C {}
+              }
+              """,
+            SourceSpec::skip),
+          java(
+            """
+              import a.*;
+              
+              public class Foo {
+                  A method(A.B ab, A.C ac) {}
+              }
+              """,
+            """
+              import a.A;
+              
+              public class Foo {
+                  A method(A.B ab, A.C ac) {}
+              }
+              """
+          )
+        );
+    }
 }
